# openapi.yaml
openapi: 3.1.0
info:
  title: Black Mesa API
  version: '1.0.0'
  description: |
    REST API for the Black Mesa Discord moderation bot.

    All endpoints under `/api/` require a Bearer JWT obtained via the
    `/api/oauth/discord` flow. The token encodes the caller's Discord user
    ID and OAuth tokens and is validated on every request.

    **Permission model**: each protected endpoint checks a specific
    `PermissionSet` flag against the caller.  Flags are a `u64` bitfield —
    see the [Permissions docs](https://docs.blackmesa.bot/bot/permissions/)
    for a full flag reference.
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `/api/oauth/discord`.  Pass as `Authorization: Bearer <token>`.
  schemas:
    Id:
      type: string
      description: Discord snowflake ID (64-bit integer encoded as a decimal string).
      pattern: '^\d+$'
      example: '123456789012345678'
...
    AuthResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Signed JWT for use in subsequent API requests.
      example:
        token: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxMjM0NTY3ODkwMTIzNDU2NzgiLCJleHAiOjE3NzUwNzIwMDB9.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
    ApiError:
      description: Error payload returned by error handlers.
      oneOf:
        - type: string
        - type: object
          additionalProperties: true
    PermissionBits:
      type: integer
      format: int64
      minimum: 0
      description: |
        64-bit permission bitfield. Each named flag is a power of two.
        Common values:
        - `MODERATION_KICK`  = 1
        - `MODERATION_BAN`   = 2
        - `MODERATION`       = 511 (all moderation flags)
        - `MUSIC`            = 130560 (all music flags)
        - `CONFIG_VIEW`      = 131072
        - `CONFIG_EDIT`      = 262144
        - `CONFIG`           = 393216
        - `INFRACTION_VIEW`  = 524288
        - `INFRACTION_EDIT`  = 1048576
        - `INFRACTION`       = 1572864
        - `UTILITY`          = 268369920
        - `ALL`              = 268435455
      example: 511
    Group:
      type: object
      required:
        - name
        - roles
        - users
        - permissions
      properties:
        name:
          type: string
          description: Unique group name within the guild.
        roles:
          type: array
          description: Discord role IDs whose holders inherit this group's permissions.
          items:
            $ref: '#/components/schemas/Id'
        users:
          type: array
          description: Individual user IDs explicitly added to this group.
          items:
            $ref: '#/components/schemas/Id'
        permissions:
          $ref: '#/components/schemas/PermissionBits'
      example:
        name: moderators
        roles:
          - '123456789012345678'
          - '222333444555666777'
        users:
          - '987654321098765432'
        permissions: 511
    PermissionOverride:
      type: object
      description: Identifies groups, roles, and users that are exempt from an automod rule.
      properties:
        groups:
          type: array
          items:
            type: string
          description: Permission group names.
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Id'
          description: Discord role IDs.
        users:
          type: array
          items:
            $ref: '#/components/schemas/Id'
          description: User IDs.
      example:
        groups:
          - moderators
        roles:
          - '123456789012345678'
        users: []
...
    CensorAction:
      type: object
      required:
        - action
        - duration
      properties:
        action:
          $ref: '#/components/schemas/InfractionType'
        duration:
          type: integer
          format: int64
          description: Duration in **milliseconds** for timed infractions (mute/ban). Ignored for kick/warn.
      example:
        action: warn
        duration: 0
    SpamAction:
      type: object
      required:
        - action
        - duration
        - threshold
      properties:
        action:
          $ref: '#/components/schemas/InfractionType'
        duration:
          type: integer
          format: int64
          description: Duration in milliseconds.
        threshold:
          type: integer
          format: int64
          description: Minimum cumulative violation count required to trigger this action.
      example:
        action: mute
        duration: 300000
        threshold: 3
    SpamInterval:
      type: object
      required:
        - interval
        - count
      properties:
        interval:
          type: integer
          format: int64
          description: Sliding window width in **milliseconds**. `0` matches all messages regardless of time.
        count:
          type: integer
          format: int64
          description: Number of events in the window required to register a hit.
      example:
        interval: 5000
        count: 5
    Censor:
      type: object
      required:
        - enabled
        - filters
        - action
      properties:
        enabled:
          type: boolean
        whitelist:
          type: boolean
          default: false
          description: When `true`, only messages that do **not** match the filters are blocked.
        filters:
          type: array
          items:
            type: string
          description: |
            Patterns to match. For `word` type: supports `*` (any suffix) and `...` (requires content
            between prefix and suffix). For `link`: exact domain names. For `invite`: invite codes.
        ignore_whitespace:
          type: boolean
          default: false
          description: Strip whitespace from message before matching.
        bypass:
          $ref: '#/components/schemas/PermissionOverride'
        action:
          $ref: '#/components/schemas/CensorAction'
      example:
        enabled: true
        filters:
          - '*badword*'
          - 'spam...'
          - 'discord.gg/abc123'
        action:
          action: warn
          duration: 0
    SpamFilter:
      type: object
      required:
        - enabled
        - filters
        - action
      properties:
        enabled:
          type: boolean
        filters:
          type: object
          description: |
            Map of `SpamType` → `SpamInterval`. Keys are `message` or `newline`.
          additionalProperties:
            $ref: '#/components/schemas/SpamInterval'
        bypass:
          $ref: '#/components/schemas/PermissionOverride'
        action:
          type: array
          description: Escalating actions at increasing violation thresholds.
          items:
            $ref: '#/components/schemas/SpamAction'
      example:
        enabled: true
        filters:
          message:
            interval: 5000
            count: 5
        action:
          - action: warn
            duration: 0
            threshold: 1
          - action: mute
            duration: 300000
            threshold: 3
    AutomodSettings:
      type: object
      required:
        - name
        - enabled
      properties:
        name:
          type: string
          description: Human-readable label for this settings block.
        enabled:
          type: boolean
        censors:
          type: object
          description: Censor rules keyed by `CensorType` (`word`, `link`, `invite`).
          additionalProperties:
            $ref: '#/components/schemas/Censor'
        spam:
          $ref: '#/components/schemas/SpamFilter'
        bypass:
          $ref: '#/components/schemas/PermissionOverride'
      example:
        name: global
        enabled: true
        censors:
          word:
            enabled: true
            filters:
              - '*badword*'
              - 'spam...'
            action:
              action: warn
              duration: 0
          invite:
            enabled: true
            filters:
              - abc123
            action:
              action: kick
              duration: 0
        spam:
          enabled: true
          filters:
            message:
              interval: 5000
              count: 5
          action:
            - action: warn
              duration: 0
              threshold: 1
            - action: mute
              duration: 300000
              threshold: 3
    Automod:
      type: object
      properties:
        global:
          oneOf:
            - $ref: '#/components/schemas/AutomodSettings'
            - type: 'null'
          description: Applies to all channels without a specific override.
        channels:
          type: object
          description: Per-channel overrides keyed by channel snowflake ID.
          additionalProperties:
            $ref: '#/components/schemas/AutomodSettings'
      example:
        global:
          name: global
          enabled: true
          censors:
            word:
              enabled: true
              filters:
                - '*badword*'
              action:
                action: warn
                duration: 0
        channels:
          '444555666777888999':
            name: media-channel
            enabled: true
            censors:
              link:
                enabled: false
                filters: []
                action:
                  action: warn
                  duration: 0
    Config:
      type: object
      required:
        - id
        - prefix
        - prefer_embeds
        - inherit_discord_perms
        - alert_on_infraction
        - send_permission_denied
        - automod_enabled
        - music_enabled
        - moderation_enabled
      properties:
        id:
          $ref: '#/components/schemas/Id'
        prefix:
          type: string
          default: '!'
          description: Command prefix used to trigger bot commands.
        mute_role:
          oneOf:
            - $ref: '#/components/schemas/Id'
            - type: 'null'
          description: Role applied by the `mute` command. Required for muting to work.
        default_warn_duration:
          type: [integer, 'null']
          format: int64
          description: Default warn expiry in **seconds** when no duration is given to `warn`.
        log_channel:
          oneOf:
            - $ref: '#/components/schemas/Id'
            - type: 'null'
          description: Channel where moderation action embeds are posted.
        prefer_embeds:
          type: boolean
          default: false
          description: Use Discord embed formatting for bot responses where supported.
        inherit_discord_perms:
          type: boolean
          default: true
          description: Map Discord role permissions (KICK_MEMBERS, BAN_MEMBERS, etc.) to Black Mesa permission flags automatically.
        alert_on_infraction:
          type: boolean
          default: true
          description: DM the target user when an infraction is issued.
        send_permission_denied:
          type: boolean
          default: true
          description: Reply with a permission-denied message when a caller lacks access.
        moderation_enabled:
          type: boolean
          default: false
          description: Enable moderation commands (`kick`, `ban`, `mute`, `warn`, etc.).
        music_enabled:
          type: boolean
          default: false
          description: Enable music/audio playback commands. Requires a Mesastream instance.
        automod_enabled:
          type: boolean
          default: false
          description: Enable the automod engine (censor and spam filters).
        permission_groups:
          description: Named permission groups. Each group binds roles/users to a `PermissionBits` value.
          oneOf:
            - type: array
              items:
                $ref: '#/components/schemas/Group'
            - type: 'null'
        automod:
          oneOf:
            - $ref: '#/components/schemas/Automod'
            - type: 'null'
          description: Automod configuration. Evaluated only when `automod_enabled` is `true`.
        command_aliases:
          description: 'Alias → command mapping, e.g. `{"w": "warn", "u": "unmute"}`.'
          oneOf:
            - type: object
              additionalProperties:
                type: string
            - type: 'null'
      example:
        id: '123456789012345678'
        prefix: '!'
        mute_role: '555666777888999000'
        default_warn_duration: 604800
        log_channel: '444555666777888999'
        prefer_embeds: true
        inherit_discord_perms: true
        alert_on_infraction: true
        send_permission_denied: true
        moderation_enabled: true
        music_enabled: false
        automod_enabled: true
        permission_groups:
          - name: moderators
            roles:
              - '222333444555666777'
            users: []
            permissions: 511
          - name: admins
            roles:
              - '333444555666777888'
            users:
              - '987654321098765432'
            permissions: 268435455
        automod:
          global:
            name: global
            enabled: true
            censors:
              word:
                enabled: true
                filters:
                  - '*badword*'
                action:
                  action: warn
                  duration: 0
            spam:
              enabled: true
              filters:
                message:
                  interval: 5000
                  count: 5
              action:
                - action: warn
                  duration: 0
                  threshold: 1
                - action: mute
                  duration: 300000
                  threshold: 3
          channels: {}
        command_aliases:
          w: warn
          u: unmute
          kb: ban
...
    InfractionType:
      type: string
      enum:
        - warn
        - mute
        - kick
        - ban
...
paths:
...
  /api/oauth/discord:
    get:
      summary: Exchange Discord OAuth code for API JWT
      description: |
        Complete the Discord OAuth2 authorization code flow and receive a signed JWT
        for use with all other API endpoints.

        Redirect your users to Discord's OAuth2 authorization URL, then send the
        returned `code` to this endpoint.
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
          description: Authorization code returned by Discord.
        - in: query
          name: redirect_uri
          required: false
          schema:
            type: string
          description: Redirect URI used in the original authorization request. Defaults to the configured value.
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxMjM0NTY3ODkwMTIzNDU2NzgiLCJleHAiOjE3NzUwNzIwMDB9.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        '400':
          description: Invalid or missing code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/oauth/refresh:
    get:
      summary: Refresh API JWT
      description: |
        Issues a fresh JWT for the currently authenticated user without re-running
        the full OAuth flow. The stored Discord credentials are preserved.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: New JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxMjM0NTY3ODkwMTIzNDU2NzgiLCJleHAiOjE3NzUwNzIwMDB9.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        '401':
          description: Token is missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/config/{id}:
    parameters:
      - in: path
        name: id
        required: true
        description: Guild snowflake ID.
        schema:
          $ref: '#/components/schemas/Id'
    get:
      summary: Get guild configuration
      description: Returns the full `Config` document for the guild. Requires `CONFIG_VIEW`.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Config document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
        '401':
          description: Unauthorized or insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Config not found (guild not set up)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      summary: Update guild configuration
      description: |
        Replaces the guild configuration. The `id` field in the body must match the path
        parameter. Requires `CONFIG_EDIT`.

        Permission groups are synced atomically: the existing groups for the guild are
        deleted and replaced with the groups provided in the request body.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Config'
            example:
              id: '123456789012345678'
              prefix: '!'
              mute_role: '555666777888999000'
              default_warn_duration: 604800
              log_channel: '444555666777888999'
              prefer_embeds: true
              inherit_discord_perms: true
              alert_on_infraction: true
              send_permission_denied: true
              moderation_enabled: true
              music_enabled: false
              automod_enabled: false
              permission_groups:
                - name: moderators
                  roles:
                    - '222333444555666777'
                  users: []
                  permissions: 511
              command_aliases:
                w: warn
                u: unmute
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
        '400':
          description: Validation error or ID mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized or insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Config not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...